pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kselect"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    # Core
    "pydantic>=2.7,<3.0",
    "numpy>=1.26,<2.0",

    # Vector index
    "faiss-cpu>=1.8.0",          # swap for faiss-gpu in [gpu] extra

    # Sparse retrieval
    "bm25s>=0.2.0",              # pure-Python BM25, no Java dep

    # Embeddings
    "sentence-transformers>=3.0,<4.0",
    "torch>=2.2,<3.0",           # pulled by sentence-transformers
    "transformers>=4.40,<5.0",

    # Chunking
    "nltk>=3.8",                 # sentence tokenizer
    "tiktoken>=0.7",             # token counting

    # Document parsing
    "pypdf>=4.0",
    "python-docx>=1.1",
    "unstructured>=0.13",        # tables, OCR dispatch

    # HTTP / async
    "httpx>=0.27",
    "anyio>=4.0",

    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # Observability
    "structlog>=24.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
gpu = [
    "faiss-gpu>=1.8.0",
]
rerank = [
    # cross-encoder already in sentence-transformers; ColBERT via:
    "colbert-ai>=0.2.19",
]
pgvector = [
    "psycopg[binary]>=3.1",      # psycopg3, async-capable
    "pgvector>=0.3",
]
pinecone = [
    "pinecone-client>=4.0",
]
chromadb = [
    "chromadb>=0.5",
]
beir = [
    "beir>=2.0",
    "datasets>=2.19",
]
eval = [
    "pandas>=2.2",
    "scipy>=1.13",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "hypothesis>=6.100",
]

[tool.hatch.build.targets.wheel]
packages = ["kselect"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
